+++ src/main/java/com/jerry/ticketing/reservation/domain/port/ReservationRepository.java
@@ -1,20 +1,21 @@
-package com.jerry.ticketing.reservation.infrastructure.repository;
+package com.jerry.ticketing.reservation.domain.port;
 
 import com.jerry.ticketing.reservation.domain.Reservation;
 import com.jerry.ticketing.reservation.domain.enums.ReservationStatus;
-import com.jerry.ticketing.seat.domain.ConcertSeat;
-import com.jerry.ticketing.seat.domain.enums.ConcertSeatStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository {
+    Reservation save(Reservation reservation);
+
+    Optional<Reservation> findById(Long id);
+
     List<Reservation> findByMemberId(Long id);
 
     Optional<Reservation> findByOrderId(String orderId);
 
     List<Reservation> findByExpiresAtBeforeAndStatus(OffsetDateTime expiresAtBefore, ReservationStatus status);
-
 }
+
